New class ShowMenu
Its purpose is to reduce the amount of code in the main method and to make the menu more scalable. The exit point was hardset to be 4 but now it will be the last number after the list of items. Later it can become a bigger menu being able to use 1 to show all available items, 2 to show the shopping card 3. to add item with a password, 4 to make an item discounted again with a password for extra security etc.