Functioning optimized program
The program is now showing messages and trying to catch errors in case of wrong input. The bug to keep adding products after they are out of stock is cleared. The percent calculation is fixed to work for all between 0 and 1 (where 1 is 100). All comments should be rewritten with a short description of the function logic and expected output.